pyproject.toml
[project]
name = "production-grade-rag-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "chromadb>=1.5.8",
    "datasets>=4.8.4",
    "faiss-cpu>=1.13.2",
    "fastapi>=0.136.0",
    "groq>=0.4.1,<1",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-core>=1.3.0",
    "langchain-groq>=1.0.0",
    "langgraph>=1.1.9",
    "mlflow>=3.11.1",
    "pydantic>=2.13.3",
    "pypdf>=6.10.2",
    "python-dotenv>=1.2.2",
    "ragas>=0.4.3",
    "requests>=2.33.1",
    "sentence-transformers>=5.4.1",
    "streamlit>=1.56.0",
    "tiktoken>=0.12.0",
    "typesense>=2.0.0",
    "uvicorn>=0.45.0",
]
